Move verifier instantiation into dedicated factory class
