+++ src/Metadata/Factory.php
@@ -2,12 +2,6 @@
 
 namespace Tuf\Metadata;
 
-use Tuf\Client\SignatureVerifier;
-use Tuf\Metadata\Verifier\RootVerifier;
-use Tuf\Metadata\Verifier\SnapshotVerifier;
-use Tuf\Metadata\Verifier\TargetsVerifier;
-use Tuf\Metadata\Verifier\TimestampVerifier;
-
 /**
  * Provides methods to load value objects for trusted metadata.
  */
@@ -21,35 +15,14 @@ class Factory
     private $storage;
 
     /**
-     * @var \DateTimeImmutable
-     */
-    private $metadataExpiration;
-
-    /**
-     * @var \Tuf\Client\SignatureVerifier
-     */
-    private $signatureVerifier;
-
-
-
-    /**
      * Factory constructor.
      *
      * @param \ArrayAccess $storage
      *   The persistent storage backend.
      */
-    public function __construct(\ArrayAccess $storage, \DateTimeImmutable $metadataExpiration)
+    public function __construct(\ArrayAccess $storage)
     {
         $this->storage = $storage;
-        $this->metadataExpiration = $metadataExpiration;
-    }
-
-    /**
-     * @param \Tuf\Client\SignatureVerifier $signatureVerifier
-     */
-    public function setSignatureVerifier(SignatureVerifier $signatureVerifier): void
-    {
-        $this->signatureVerifier = $signatureVerifier;
     }
 
     /**
@@ -84,35 +57,5 @@ class Factory
         } else {
             return null;
         }
-    }
-
-    /**
-     * Gets a metadata verifier for a role.
-     *
-     * @param string $role
-     *
-     * @return \Tuf\Metadata\Verifier\RootVerifier|\Tuf\Metadata\Verifier\SnapshotVerifier|\Tuf\Metadata\Verifier\TargetsVerifier|\Tuf\Metadata\Verifier\TimestampVerifier
-     */
-    public function getVerifier(string $role)
-    {
-        $trustedMetadata = $this->load($role);
-        switch ($role) {
-            case RootMetadata::TYPE:
-                $verifier = new RootVerifier($this->signatureVerifier, $this->metadataExpiration, $trustedMetadata);
-                break;
-            case SnapshotMetadata::TYPE:
-                /** @var \Tuf\Metadata\TimestampMetadata $timestampMetadata */
-                $timestampMetadata = $this->load(TimestampMetadata::TYPE);
-                $verifier = new SnapshotVerifier($this->signatureVerifier, $this->metadataExpiration, $trustedMetadata, $timestampMetadata);
-                break;
-            case TimestampMetadata::TYPE:
-                $verifier = new TimestampVerifier($this->signatureVerifier, $this->metadataExpiration, $trustedMetadata);
-                break;
-            default:
-                /** @var \Tuf\Metadata\SnapshotMetadata $snapshotMetadata */
-                $snapshotMetadata = $this->load(SnapshotMetadata::TYPE);
-                $verifier = new TargetsVerifier($this->signatureVerifier, $this->metadataExpiration, $trustedMetadata, $snapshotMetadata);
-        }
-        return $verifier;
     }
 }
